% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utilities.R
\name{suq2.clustering.gldClustering}
\alias{suq2.clustering.gldClustering}
\title{Clustering of the GLDs in function of its l2, l3 and l4 values}
\usage{
suq2.clustering.gldClustering(lambdas, no_clusters, l234 = TRUE)
}
\arguments{
\item{lambdas}{A matrix of n x m x 4 of all the lambda values.}

\item{no_clusters}{Number of clusters.}
}
\description{
TODO.
}
\examples{

suq2.clustering.gldClustering(lambdas, 10, TRUE)

}
\author{
Noel Moreno Lemus
}
