% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_fit.R
\name{suq2.fit.gldfitLmoments}
\alias{suq2.fit.gldfitLmoments}
\title{Compute the lambda values of the GLD using the method of the Lmoments}
\usage{
suq2.fit.gldfitLmoments(data)
}
\arguments{
\item{data}{The raw data to fit.}
}
\description{
This function allows you to show image with scale.
}
\examples{
suq2.fit.gldfitLmoments(data)
}
\keyword{lmoments}
