% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld.plot.R, R/plot.gld.R
\name{suq2.plot.gld}
\alias{suq2.plot.gld}
\alias{suq2.plot.gld}
\title{Plot the GLD based in its lambda values.}
\usage{
suq2.plot.gld(n = 1000, L = c(0, 2, 4, 4), param = "fmkl")

suq2.plot.gld(n = 1000, L = c(0, 2, 4, 4), param = "fmkl")
}
\arguments{
\item{n}{the directory where the .csv files are stored.}

\item{L}{a pattern to read the .csv (e.g. pattern="^[h]").}

\item{param}{(default = "fmkl")}

\item{n}{the directory where the .csv files are stored.}

\item{L}{a pattern to read the .csv (e.g. pattern="^[h]").}

\item{param}{(default = "fmkl")}
}
\description{
This function read all the csv that are stored in a directory and create an NxMxS array, where N and M are sparial dimensions
and S is the number of simulations on each spatial point.

This function read all the csv that are stored in a directory and create an NxMxS array, where N and M are sparial dimensions
and S is the number of simulations on each spatial point.
}
\examples{
suq2.plot.gld(L = c(0, 2, 0.14, 0.14))

suq2.plot.gld(L = c(0, 2, 0.14, 0.14))

}
\author{
Noel Moreno Lemus

Noel Moreno Lemus
}
