% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utilities.R
\name{suq2.plot.gldClustersL3L4}
\alias{suq2.plot.gldClustersL3L4}
\title{Plot the clusters in l3-l4 space.}
\usage{
suq2.plot.gldClustersL3L4(clusters, x)
}
\arguments{
\item{clusters}{An n x m matrix wiht the clusters by positions.}

\item{x}{Array used to create the clusters in function @method gldClustering.}
}
\description{
TODO.
}
\examples{

suq2.plot.gldClustersL3L4(clusters, x)

}
\author{
Noel Moreno Lemus
}
