% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utilities.R
\name{suq2.plot.gldToPlot}
\alias{suq2.plot.gldToPlot}
\title{Return (x, y) values of the GLD based in the Lambda values}
\usage{
suq2.plot.gldToPlot(L, param = "fmkl")
}
\arguments{
\item{L}{Lambda values of the GLD.}
}
\description{
This function allows you to return the (x,y) values of the GLD, based in the Lambda values.
}
\examples{
L = c(0, 2, 0.25, 1.5)
suq2.plot.gldToPlot(L)

}
