% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_display.R
\name{suq2.plot.image_display}
\alias{suq2.plot.image_display}
\title{Color image dislay with scale}
\usage{
suq2.plot.image_display(data)
}
\arguments{
\item{date}{The image matrix you like to display.}
}
\description{
This function allows you to show image with scale.
}
\examples{
suq2.plot.image_display()
}
\keyword{image}
