% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gld_validity.R
\name{suq2.utils.checkGLDValid1}
\alias{suq2.utils.checkGLDValid1}
\title{Check if one GLD is valid}
\usage{
suq2.utils.checkGLDValid1(lambdas)
}
\arguments{
\item{lambdas}{lambda values of a single GLD function.}
}
\description{
This function check if one GLD is valid.
}
\examples{
suq2.utils.checkGLDValid1(lambdas)
}
\keyword{lambdas}
