% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utilities.R
\name{suq2.utils.distGLDComparison}
\alias{suq2.utils.distGLDComparison}
\title{Function to be used as a distance function in clustering algoritms}
\usage{
suq2.utils.distGLDComparison(x, centers)
}
\arguments{
\item{x}{dataset.}

\item{centers}{the centroids to be analized.}
}
\description{
This function return the distances between all the centroids and the elements of the dataset, using a KS-test as a measure of the distance.
}
\examples{
TODO

}
\author{
Noel Moreno Lemus
}
